requirements.txt
pypdf>=4.0.0
tqdm>=4.66.0
google-generativeai>=0.7.0
chromadb>=0.4.0
